requirements.txt
fastapi
uvicorn
motor
pydantic
pymongo
pymongo[srv]
pydantic[email]
PyJWT
python-decouple
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[dotenv]
